requirements.txt
# local package
-e .

# external requirements (please freeze version to the one you are using)
bandit==1.7.1
black==21.12b0
doit==0.34.0
dvc==2.9.3
flake8==4.0.1
loguru==0.5.3
mlflow==1.22.0
pylint==2.12.2
pytest==6.2.5
pytest-cov==3.0.0
python-dotenv==0.19.2

# Add new requirements below
joblib==1.0.1
jupyterlab==3.2.8
lightgbm==3.3.1
numpy==1.19.5
matplotlib==3.5.1
pycaret==2.3.6
pydrive2==1.10.0
pyyaml==5.4.1
scikit-learn==0.23.2
tokenize-rt==4.2.1